+++ src/glob/fs/LazyEntry.d.ts
@@ -1,0 +1,3 @@
+import { type LazyDir } from './LazyDir.js'
+import { type LazyFile } from './LazyFile.js'
+export type LazyEntry = LazyFile | LazyDir
