+++ src/glob/glob.js
@@ -1,0 +1,3 @@
+import { LazyGlob } from './LazyGlob.js'
+
+export const glob = new LazyGlob()
