+++ src/glob/matchers/RootMatcher.js
@@ -1,0 +1,19 @@
+/** @implements {Matcher} */
+export class RootMatcher {
+  /**
+   * @type {boolean}
+   * @readonly
+   */
+  negating = false
+
+  /** @type {Matcher[]} */
+  next = []
+  /**
+   * @param {import("../fs/LazyEntry.js").LazyEntry} _entry
+   * @param {MatchOptions} _options
+   * @return {MatchResult}
+   */
+  match(_entry, _options) {
+    return 'partial'
+  }
+}
